feat: add force bun install feat: option worker runner refactor: use @nx/js
