+++ packages/nx-bun/src/generators/init/generator.spec.ts
@@ -1,7 +1,7 @@
 import { createTreeWithEmptyWorkspace } from '@nx/devkit/testing';
 import { Tree, readProjectConfiguration } from '@nx/devkit';
 
-import { initGenerator } from './generator';
+import { initGenerator } from './init';
 import { InitGeneratorSchema } from './schema';
 
 describe('init generator', () => {
